Revert "Add module"
