+++ CMakeLists.txt
@@ -2,9 +2,9 @@ cmake_minimum_required(VERSION 3.7)
 
 project(BinaryMapper)
 
-set(CMAKE_CXX_STANDARD 17)
-
 include(cmake/make.cmake)
+
+set(CMAKE_CXX_STANDARD 17)
 
 set(CMAKE_INCLUDE_CURRENT_DIR ON)
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -g -Wall")
@@ -12,6 +12,5 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -g -Wall")
 include_directories(${CMAKE_SOURCE_DIR}/modules/common/include)
 
 static_library(${PROJECT_NAME} SOURCES "src/")
-clang_tidy_analysis(${PROJECT_NAME})
 
-add_subdirectory(examples)
+add_subdirectory(example)
