--- examples/driver_example/lib/test.c
@@ -1,4 +1,0 @@
-void test()
-{
-    
-}
