--- examples/py_example/py_example.py
@@ -1,6 +1,0 @@
-#!/usr/bin/python
-
-import sys
-
-print('Py Hello, world!')
-print(sys.argv[1])
