--- examples/sh_example/sh_example.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-echo Sh Hello world!
-echo $1
