--- include/ModuleObject.h
@@ -1,1 +1,0 @@
-#include "../src/ModuleObject.h"
