--- src/ExecType.h
@@ -1,8 +1,0 @@
-#pragma once
-
-enum class ExecType
-{
-    Unknown,
-    LocalExec,
-    LoadExec
-};
